ai-service/requirements.txt
google-genai>=0.8.0
langchain>=0.3.0
langchain-google-genai>=0.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.20.0
httpx>=0.27.0
pydantic>=2.0.0
